Added permission on form

+++ FinVentoryAPI/Controllers/LocationController.cs
@@ -1,40 +1,55 @@
 ﻿using FinVentoryAPI.DTOs.CompanyDTOs;
 using FinVentoryAPI.DTOs.LocationDTOs;
+using FinVentoryAPI.DTOs.RoleRightsDTOs;
 using FinVentoryAPI.Services.Implementations;
 using FinVentoryAPI.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore.Metadata;
 
 namespace FinVentoryAPI.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class LocationController : ControllerBase
     {
         private readonly ILocationService _locationService;
+        private readonly IRoleRightService _roleRightService;
 
-        public LocationController(ILocationService locationService)
+        public LocationController(ILocationService locationService, IRoleRightService roleRightService)
         {
             _locationService = locationService;
+            _roleRightService = roleRightService;
+
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateLocationDTO dto)
         {
+            var permission = await GetPermission("Add");
+
+            if (permission == null || !permission.CanAdd)
+                return Forbid("No Add Permission");
+
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            //int userId = 1; // Later get from JWT claim
-
             var result = await _locationService.CreateAsync(dto);
 
-            return CreatedAtAction(nameof(GetById), new { id = result.CompanyId }, result);
+            return CreatedAtAction(nameof(GetById),
+                new { id = result.LocationId }, result);
         }
 
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            var permission = await GetPermission("View");
+
+            if (permission == null || !permission.CanView)
+                return Forbid("No View Permission");
+
             var locations = await _locationService.GetAllAsync();
             return Ok(locations);
         }
@@ -42,6 +57,11 @@ namespace FinVentoryAPI.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(int id)
         {
+            var permission = await GetPermission("View");
+
+            if (permission == null || !permission.CanView)
+                return Forbid("No View Permission");
+
             var location = await _locationService.GetByIdAsync(id);
 
             if (location == null)
@@ -52,10 +72,13 @@ namespace FinVentoryAPI.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(int id, [FromBody] UpdateLocationDTO dto)
         {
+            var permission = await GetPermission("Edit");
+
+            if (permission == null || !permission.CanEdit)
+                return Forbid("No Edit Permission");
+
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
-
-          //  int userId = 1; // Later get from JWT
 
             var updated = await _locationService.UpdateAsync(id, dto);
 
@@ -68,7 +91,12 @@ namespace FinVentoryAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
-            int userId = 1; // Later get from JWT
+            var permission = await GetPermission("Delete");
+
+            if (permission == null || !permission.CanDelete)
+                return Forbid("No Delete Permission");
+
+            var userId = Convert.ToInt32(User.FindFirst("UserId")?.Value);
 
             var deleted = await _locationService.DeleteAsync(id, userId);
 
@@ -76,6 +104,27 @@ namespace FinVentoryAPI.Controllers
                 return NotFound(new { message = "Location not found" });
 
             return Ok(new { message = "Location deleted successfully" });
+        }
+        
+        private async Task<FormPermissionDto?> GetPermission(string actionType)
+        {
+            // 🔹 Get RoleId from JWT
+            var roleIdClaim = User.FindFirst("RoleId")?.Value;
+            if (roleIdClaim == null)
+                return null;
+
+            int roleId = Convert.ToInt32(roleIdClaim);
+
+            // 🔹 Get MenuItemId from Header
+            if (!Request.Headers.TryGetValue("MenuItemId", out var menuHeader))
+                return null;
+
+            int menuItemId = Convert.ToInt32(menuHeader);
+
+            var permission = await _roleRightService
+                .GetFormPermissionsAsync(menuItemId, roleId);
+
+            return permission;
         }
     }
 }
